+++ src/domain/exercises/index.ts
@@ -1,5 +1,6 @@
 export * from './generators';
 export * from './grading';
+export * from './modes';
 export * from './speech';
 export * from './types';
 export * from './variations';